Transfer to redux, complete redux for userinfo
