--- src/actions/index.js
@@ -1,9 +1,0 @@
-// ACTION TYPES
-export const DATA = 'DATA'
-
-// ACTION CREATORS
-export const createData = () => {
-  return {
-    type: DATA
-  }
-}
