+++ src/actions/types.js
@@ -1,0 +1,2 @@
+export const LOGIN = 'LOGIN'
+export const GET_USER = 'GET_USER'
