+++ src/clientinfo.js
@@ -1,2 +1,2 @@
 export const CLIENT_ID='sv7qodeq1ii9rnpnk4jzs5tzmvh036'
-export const REDIRECT_URI='http://0e75100e.ngrok.io'
+export const REDIRECT_URI='http://b295a871.ngrok.io'
