--- src/reducers/dataReducer.js
@@ -1,10 +1,0 @@
-import { DATA } from '../actions/'
-
-const dataReducer = (state, action) => {
-  switch(action.type) {
-    case DATA:
-    
-  }
-}
-
-export default dataReducer;
