+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import userInfoReducers from './userInfoReducer';
+
+export default combineReducers({
+  user: userInfoReducers
+})
